ml_service/requirements.txt
flask
flask-cors
ultralytics
pillow
requests
